genisoimage.1 cleanup.  Step 1: argument quoting.
If a nroff "" string extends to the end of a line, the closing " is
often omitted.  Although this is legal, it is also confusing, so add
closing " where needed so that all sets of "" are matched.

Also, .B and .I can take multiple arguments, so remove quotes from
lines like .B "CHARACTER SETS".


git-svn-id: svn://svn.debian.org/debburn/cdrkit/trunk@613 a95a6be8-091b-0410-adaf-d31e6857962f