genisoimage.1 cleanup.  Step 3: quotes inside the text (not in nroff cmds).
Make ""  ''  `' a bit more consistent.  Now `' is used for quoting single
characters like `#'.  Longer strings that need to be quoted are
occasionally done with "" but usually just changed to italics or
omitted ("El Torito" does not need quoting, for example).  Note: we
could use \[lq]\[rq] instead of "", but that's a bit anal.

* Yes, I _do_ know that `' is incorrect for ASCII text.  But in nroff,
  it is rendered the same as \[oq]\[cq] (left and right single quotes).

Other drive-by cleanups:
- Replace .LP with .PP (they are synonyms)
- Replace "it's" with "its" except the one spot where "it's" is correct
- Use .IP to fix some indentation in the EXAMPLES section
- Hyphenate "7-bit" and friends.


git-svn-id: svn://svn.debian.org/debburn/cdrkit/trunk@615 a95a6be8-091b-0410-adaf-d31e6857962f