genisoimage.1 cleanup.  Step 4: random typography.
Render "\" using "\(rs" rather than "\\", since \\ is ambiguous in some
situations (rendering "\r" is where I noticed this).

Replace hyphens ("-") with em dashes ("\(em").  An em dash is used as a
separator in the middle of a sentence - like this.  Em dashes look very
different from hyphens in typeset output (groff -Tps): "\(em" is quite
long, "-" is quite short.  The ASCII dash "\-" is in between.


git-svn-id: svn://svn.debian.org/debburn/cdrkit/trunk@616 a95a6be8-091b-0410-adaf-d31e6857962f